fix: dont drain source while drain in progress
Misbehaving streams can end up calling drain multiple times while
a drain is in progress. This is bad news for pull streams as it
breaks the contract and you end up with weirdness like end of stream
events getting triggered before data event callbacks are fired.

While tracking this down I applied a similar fix to pull-file-reader
but that modules shouldn't have to guard against bad pullers.

Then I found https://github.com/pull-stream/pull-stream-to-stream/pull/6
which deals with the same issue in pull-stream-to-steam (thanks @mmckegg!)
and recreated it here. The test on this PR is copied from there.
The gist is, as this module operates at the boundary of node-streams
to pull-streams, it should deal with bad node-stream behaviour to
avoid messing up connected pull-streams.

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>